ticket_booking_crud/requirements.txt
# Core FastAPI Dependencies
fastapi==0.104.1              # Modern, fast web framework for building APIs
uvicorn[standard]==0.24.0     # ASGI server for running FastAPI applications

# Database Dependencies
sqlalchemy==2.0.23            # Python SQL toolkit and ORM
alembic==1.12.1              # Database migration tool for SQLAlchemy
psycopg2-binary==2.9.9       # PostgreSQL database adapter
# sqlite3 is built into Python - no need to install

# Data Validation & Serialization
pydantic==2.5.0              # Data validation using Python type hints
pydantic-settings==2.0.3     # Settings management for Pydantic

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT token handling
passlib[bcrypt]==1.7.4       # Password hashing utilities
python-multipart==0.0.6      # Form data parsing for file uploads

# Environment & Configuration
python-dotenv==1.0.0         # Load environment variables from .env file

# Testing Dependencies
pytest==7.4.3               # Testing framework
pytest-asyncio==0.21.1      # Testing async code
httpx==0.25.2               # HTTP client for testing FastAPI

# Development Dependencies
black==23.11.0              # Code formatter
flake8==6.1.0               # Linting tool
mypy==1.7.1                 # Static type checker

# Additional Utilities
email-validator==2.1.0      # Email validation
python-slugify==8.0.1       # URL-friendly string generation 